Added tests for loader func
